# docs/swagger.yaml
basePath: /api
definitions:
...
  GAMERS-BE_internal_comment_application_dto.AuthorResponse:
    properties:
      avatar:
        type: string
      tag:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  GAMERS-BE_internal_comment_application_dto.CommentResponse:
    properties:
      author:
        $ref: '#/definitions/GAMERS-BE_internal_comment_application_dto.AuthorResponse'
      comment_id:
        type: integer
      content:
        type: string
      contest_id:
        type: integer
      created_at:
        type: string
      modified_at:
        type: string
    type: object
  GAMERS-BE_internal_comment_application_dto.CreateCommentRequest:
    properties:
      content:
        maxLength: 255
        type: string
    required:
    - content
    type: object
  GAMERS-BE_internal_comment_application_dto.UpdateCommentRequest:
    properties:
      content:
        maxLength: 255
        type: string
    required:
    - content
    type: object
...
  GAMERS-BE_internal_global_common_dto.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  GAMERS-BE_internal_global_response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gamers.com
    name: API Support
  description: API Server for GAMERS platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GAMERS API
  version: "1.0"
paths:
...
  /api/contests/{id}/comments:
    get:
      consumes:
      - application/json
      description: Get all comments for a contest with pagination
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - default: created_at
        description: Sort field (created_at, modified_at)
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_global_common_dto.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get comments for a contest
      tags:
      - contest-comments
    post:
      consumes:
      - application/json
      description: Create a new comment for a contest
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment creation request
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_comment_application_dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_comment_application_dto.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Create a new comment
      tags:
      - contest-comments
  /api/contests/{id}/comments/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete a comment (owner only)
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - contest-comments
    get:
      consumes:
      - application/json
      description: Get a specific comment by its ID
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_comment_application_dto.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get a comment by ID
      tags:
      - contest-comments
    patch:
      consumes:
      - application/json
      description: Update a comment (owner only)
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: integer
      - description: Comment update request
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_comment_application_dto.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_comment_application_dto.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - contest-comments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"